requirements.txt
Django==1.7
django-tastypie==0.12.2
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
